Remove variable-length arrays.
Variable-length arrays cause trouble for certain compilers
(*cough* MSVC *cough*), and this patch removes them. Most of the
VLAs were in the functions celt_fir and celt_iir, which were unused.
The final VLA was in _celt_autocorr, and this patch switches to
using a buffer passed as a parameter.