fixes for non-avx targets
when emulating avx for non-avx targets, the _m256i type is being typedef'd,
however, this type is defined in avxintrin.h which is only included via
immintrin.h if avx is enabled.  I've added an inline definition that seems
to work on my x86_64 machine with clang when __AVX__ is not defined.

Also, mm256_maddubs_epi16 was unable to compile for non-avx targets,
I've addressed this by only defining this function when either __AVX2__
or __SSSE3__ is defined.